- Refactor maven pom hierarchy and package names involved in shardingsphere. - Added ShardingSphere JDBC Driver integration example. - This adjustment is intended to emphasize definitions and avoid unreasonable ambiguities with the work being done on the ShardingSphere R2DBC Driver.
