+++ third-part-samples/pom.xml
@@ -8,13 +8,10 @@
         <version>1.0.0</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
     <artifactId>third-part-samples</artifactId>
     <packaging>pom</packaging>
-
     <modules>
         <module>quartz-sample</module>
-        <module>sharding-jdbc-sample</module>
-        <module>shardingSphere-jdbc-sample</module>
+        <module>shardingsphere-sample</module>
     </modules>
 </project>
