--- third-part-samples/sharding-jdbc-sample/src/main/java/com/baomidou/samples/sharding/ShardingApplication.java
@@ -1,29 +1,0 @@
-/*
- * Copyright © ${project.inceptionYear} organization baomidou
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.baomidou.samples.sharding;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-@MapperScan("com.baomidou.samples.sharding.mapper")
-public class ShardingApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ShardingApplication.class, args);
-    }
-}
