--- third-part-samples/sharding-jdbc-sample/src/main/java/com/baomidou/samples/sharding/config/MyDataSourceConfiguration.java
@@ -1,82 +1,0 @@
-/*
- * Copyright © ${project.inceptionYear} organization baomidou
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.baomidou.samples.sharding.config;
-
-import com.baomidou.dynamic.datasource.DynamicRoutingDataSource;
-import com.baomidou.dynamic.datasource.provider.AbstractDataSourceProvider;
-import com.baomidou.dynamic.datasource.provider.DynamicDataSourceProvider;
-import com.baomidou.dynamic.datasource.spring.boot.autoconfigure.DynamicDataSourceProperties;
-import org.apache.shardingsphere.shardingjdbc.jdbc.core.datasource.MasterSlaveDataSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-
-import javax.annotation.Resource;
-import javax.sql.DataSource;
-import java.util.HashMap;
-import java.util.Map;
-
-@Configuration
-public class MyDataSourceConfiguration {
-
-    @Resource
-    private DynamicDataSourceProperties properties;
-
-    /**
-     * 分片: shardingDataSource
-     * 主从: masterSlaveDataSource
-     * 根据自己场景修改注入
-     */
-//    @Lazy 某些springBoot版本不加会报错（暂不清楚原理0 0）
-    @Resource
-    private MasterSlaveDataSource masterSlaveDataSource;
-    //    @Lazy
-//    @Resource
-//    private ShardingDataSource shardingDataSource;
-
-    @Bean
-    public DynamicDataSourceProvider dynamicDataSourceProvider() {
-        return new AbstractDataSourceProvider() {
-
-            @Override
-            public Map<String, DataSource> loadDataSources() {
-                Map<String, DataSource> dataSourceMap = new HashMap<>();
-                dataSourceMap.put("sharding", masterSlaveDataSource);
-                //下面的代码可以把 shardingJdbc 内部管理的子数据源也同时添加到动态数据源里 (根据自己需要选择开启+注解了@Lazy被代理了不可以)
-                Map<String, DataSource> shardingInnerDataSources = masterSlaveDataSource.getDataSourceMap();
-                dataSourceMap.putAll(shardingInnerDataSources);
-                return dataSourceMap;
-            }
-        };
-    }
-
-    /**
-     * 将动态数据源设置为首选的
-     * 当spring存在多个数据源时, 自动注入的是首选的对象
-     * 设置为主要的数据源之后，就可以支持shardingJdbc原生的配置方式了
-     */
-    @Primary
-    @Bean
-    public DataSource dataSource() {
-        DynamicRoutingDataSource dataSource = new DynamicRoutingDataSource();
-        dataSource.setPrimary(properties.getPrimary());
-        dataSource.setStrict(properties.getStrict());
-        dataSource.setStrategy(properties.getStrategy());
-        dataSource.setP6spy(properties.getP6spy());
-        dataSource.setSeata(properties.getSeata());
-        return dataSource;
-    }
-}
