--- third-part-samples/sharding-jdbc-sample/src/main/java/com/baomidou/samples/sharding/entity/User.java
@@ -1,28 +1,0 @@
-/*
- * Copyright © ${project.inceptionYear} organization baomidou
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.baomidou.samples.sharding.entity;
-
-import lombok.Data;
-
-@Data
-public class User {
-
-    private Integer id;
-
-    private String name;
-
-    private Integer age;
-}
