--- third-part-samples/sharding-jdbc-sample/src/main/java/com/baomidou/samples/sharding/service/impl/UserServiceImpl.java
@@ -1,57 +1,0 @@
-/*
- * Copyright © ${project.inceptionYear} organization baomidou
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.baomidou.samples.sharding.service.impl;
-
-
-import com.baomidou.dynamic.datasource.annotation.DS;
-import com.baomidou.samples.sharding.entity.User;
-import com.baomidou.samples.sharding.mapper.UserMapper;
-import com.baomidou.samples.sharding.service.UserService;
-import org.springframework.stereotype.Service;
-
-import javax.annotation.Resource;
-import java.util.List;
-
-@Service
-public class UserServiceImpl implements UserService {
-
-    @Resource
-    private UserMapper userMapper;
-
-    @Override
-    public List<User> selectUsersFromMaster() {
-        return userMapper.selectUsers();
-    }
-
-
-    @Override
-    @DS("sharding")
-    public List<User> selectUsersFromShardingSlave() {
-        return userMapper.selectUsers();
-    }
-
-    @DS("sharding")
-    @Override
-    public void addUser(User user) {
-        userMapper.addUser(user.getName(), user.getAge());
-    }
-
-    @DS("sharding")
-    @Override
-    public void deleteUserById(Long id) {
-        userMapper.deleteUserById(id);
-    }
-}
