+++ third-part-samples/shardingsphere-sample/pom.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>third-part-samples</artifactId>
+        <groupId>com.baomidou</groupId>
+        <version>1.0.0</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+    <artifactId>shardingsphere-sample</artifactId>
+    <packaging>pom</packaging>
+    <modules>
+        <module>shardingsphere-jdbc-5.x-spring-sample</module>
+        <module>shardingsphere-jdbc-4.x-spring-sample</module>
+        <module>shardingsphere-jdbc-5.x-core-sample</module>
+    </modules>
+</project>
