+++ third-part-samples/shardingsphere-sample/shardingsphere-jdbc-5.x-core-sample/src/main/resources/application.yml
@@ -1,0 +1,21 @@
+spring:
+  datasource:
+    dynamic:
+      datasource:
+        master:
+          username: sa
+          password: ""
+          url: jdbc:h2:mem:master;MODE=MYSQL;DATABASE_TO_LOWER=TRUE;INIT=RUNSCRIPT FROM 'classpath:db/schema.sql'
+          driver-class-name: org.h2.Driver
+        test:
+          username: sa
+          password: ""
+          url: jdbc:h2:mem:test;MODE=MYSQL;DATABASE_TO_LOWER=TRUE;INIT=RUNSCRIPT FROM 'classpath:db/schema.sql'
+          driver-class-name: org.h2.Driver
+        shardingSphere:
+          url: jdbc:shardingsphere:classpath:config.yaml
+          driver-class-name: org.apache.shardingsphere.driver.ShardingSphereDriver
+logging:
+  level:
+    com.baomidou: debug
+    org.apache.shardingsphere: debug
