+++ third-part-samples/shardingsphere-sample/shardingsphere-jdbc-5.x-spring-sample/src/main/java/com/baomidou/samples/shardingSphere/jdbc/v5/spring/controller/UserController.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright © ${project.inceptionYear} organization baomidou
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.baomidou.samples.shardingSphere.jdbc.v5.spring.controller;
+
+import com.baomidou.samples.shardingSphere.jdbc.v5.spring.entity.User;
+import com.baomidou.samples.shardingSphere.jdbc.v5.spring.service.UserService;
+import lombok.AllArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.Random;
+
+@RestController
+@AllArgsConstructor
+@RequestMapping("/users")
+public class UserController {
+    private static final Random RANDOM = new Random();
+    private final UserService userService;
+
+    @GetMapping("master")
+    public List<User> master() {
+        return userService.selectUsersFromMaster();
+    }
+
+    @GetMapping("sharding_sphere")
+    public List<User> shardingSlave() {
+        return userService.selectUsersFromShardingSlave();
+    }
+
+    @PostMapping("sharding_sphere")
+    public User addUser() {
+        User user = new User();
+        user.setName("测试用户" + RANDOM.nextInt());
+        user.setAge(RANDOM.nextInt(100));
+        userService.addUser(user);
+        return user;
+    }
+
+    @DeleteMapping("sharding_sphere/{id}")
+    public String deleteUser(@PathVariable Long id) {
+        userService.deleteUserById(id);
+        return "成功删除用户" + id;
+    }
+}
