+++ third-part-samples/shardingsphere-sample/shardingsphere-jdbc-5.x-spring-sample/src/main/java/com/baomidou/samples/shardingSphere/jdbc/v5/spring/service/impl/UserServiceImpl.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright © ${project.inceptionYear} organization baomidou
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.baomidou.samples.shardingSphere.jdbc.v5.spring.service.impl;
+
+import com.baomidou.dynamic.datasource.annotation.DS;
+import com.baomidou.samples.shardingSphere.jdbc.v5.spring.entity.User;
+import com.baomidou.samples.shardingSphere.jdbc.v5.spring.service.UserService;
+import com.baomidou.samples.shardingSphere.jdbc.v5.spring.mapper.UserMapper;
+import org.springframework.stereotype.Service;
+
+import javax.annotation.Resource;
+import java.util.List;
+
+@Service
+public class UserServiceImpl implements UserService {
+    @Resource
+    private UserMapper userMapper;
+
+    @Override
+    public List<User> selectUsersFromMaster() {
+        return userMapper.selectUsers();
+    }
+
+    @Override
+    @DS("shardingSphere")
+    public List<User> selectUsersFromShardingSlave() {
+        return userMapper.selectUsers();
+    }
+
+    @DS("shardingSphere")
+    @Override
+    public void addUser(User user) {
+        userMapper.addUser(user.getName(), user.getAge());
+    }
+
+    @DS("shardingSphere")
+    @Override
+    public void deleteUserById(Long id) {
+        userMapper.deleteUserById(id);
+    }
+}
