+++ third-part-samples/shardingsphere-sample/shardingsphere-jdbc-5.x-spring-sample/src/main/resources/application.yml
@@ -1,0 +1,74 @@
+spring:
+  shardingsphere:
+    datasource:
+      names: shardingmaster,shardingslave0,shardingslave1
+      shardingmaster:
+        type: com.zaxxer.hikari.HikariDataSource
+        driver-class-name: org.h2.Driver
+        jdbc-url: jdbc:h2:mem:master;MODE=MYSQL;DATABASE_TO_LOWER=TRUE;INIT=RUNSCRIPT FROM 'classpath:db/schema.sql'
+        username: sa
+        password: ""
+      shardingslave0:
+        type: com.zaxxer.hikari.HikariDataSource
+        driver-class-name: org.h2.Driver
+        jdbc-url: jdbc:h2:mem:slave1;MODE=MYSQL;DATABASE_TO_LOWER=TRUE;INIT=RUNSCRIPT FROM 'classpath:db/schema.sql'
+        username: sa
+        password: ""
+      shardingslave1:
+        type: com.zaxxer.hikari.HikariDataSource
+        driver-class-name: org.h2.Driver
+        jdbc-url: jdbc:h2:mem:slave2;MODE=MYSQL;DATABASE_TO_LOWER=TRUE;INIT=RUNSCRIPT FROM 'classpath:db/schema.sql'
+        username: sa
+        password: ""
+    rules:
+      sharding:
+        tables:
+          t_order:
+            actualDataNodes: shardingmaster.t_order$->{0..1}
+            tableStrategy:
+              standard:
+                shardingColumn: order_id
+                shardingAlgorithmName: baomidou-inline
+            keyGenerateStrategy:
+              column: order_id
+              keyGeneratorName: baomidou-snowflake
+        sharding-algorithms:
+          baomidou-inline:
+            type: INLINE
+            props:
+              algorithm-expression: t_order$->{order_id % 2}
+              allow-range-query-with-inline-sharding: true
+        key-generators:
+          baomidou-snowflake:
+            type: SNOWFLAKE
+            props:
+              max-vibration-offset: 1
+              max-tolerate-time-difference-milliseconds: 10
+      readwrite-splitting:
+        data-sources:
+          baomidou-readwrite-data-sources:
+            type: Static
+            props:
+              write-data-source-name: shardingmaster
+              read-data-source-names: shardingslave0,shardingslave1
+            load-balancer-name: baomidou-load-balance-algorithm
+        load-balancers:
+          baomidou-load-balance-algorithm:
+            type: ROUND_ROBIN
+  datasource:
+    dynamic:
+      datasource:
+        master:
+          username: sa
+          password: ""
+          url: jdbc:h2:mem:master;MODE=MYSQL;DATABASE_TO_LOWER=TRUE;INIT=RUNSCRIPT FROM 'classpath:db/schema.sql'
+          driver-class-name: org.h2.Driver
+        test:
+          username: sa
+          password: ""
+          url: jdbc:h2:mem:test;MODE=MYSQL;DATABASE_TO_LOWER=TRUE;INIT=RUNSCRIPT FROM 'classpath:db/schema.sql'
+          driver-class-name: org.h2.Driver
+logging:
+  level:
+    com.baomidou: debug
+    org.apache.shardingsphere: debug
